Finished Pascal and added GCD snippet

+++ Python/Snippets/Primality.py
@@ -9,7 +9,7 @@ def isPrime(n):
     if n % 2 == 0 or n % 3 == 0:
         return False
 
-    for i in range(5, math.sqrt(n), 6):
+    for i in range(5, int(math.sqrt(n)), 6):
         if n % i == 0 or n % (i + 2) == 0:
             return False
